ASC.Data.Backup.BackgroundTasks: refactoring
