+++ common/services/ASC.Data.Backup.BackgroundTasks/GlobalUsings.cs
@@ -54,3 +54,4 @@ global using Microsoft.Extensions.Hosting.WindowsServices;
 global using Microsoft.Extensions.Options;
 
 global using Newtonsoft.Json;
+global using ASC.EventBus.Events;
