--- common/services/ASC.Data.Backup.BackgroundTasks/IntegrationEvents/EventHandling/BackupRequestIntegrationEventHandler.cs
@@ -1,91 +1,0 @@
-﻿// (c) Copyright Ascensio System SIA 2010-2022
-//
-// This program is a free software product.
-// You can redistribute it and/or modify it under the terms
-// of the GNU Affero General Public License (AGPL) version 3 as published by the Free Software
-// Foundation. In accordance with Section 7(a) of the GNU AGPL its Section 15 shall be amended
-// to the effect that Ascensio System SIA expressly excludes the warranty of non-infringement of
-// any third-party rights.
-//
-// This program is distributed WITHOUT ANY WARRANTY, without even the implied warranty
-// of MERCHANTABILITY or FITNESS FOR A PARTICULAR  PURPOSE. For details, see
-// the GNU AGPL at: http://www.gnu.org/licenses/agpl-3.0.html
-//
-// You can contact Ascensio System SIA at Lubanas st. 125a-25, Riga, Latvia, EU, LV-1021.
-//
-// The  interactive user interfaces in modified source and object code versions of the Program must
-// display Appropriate Legal Notices, as required under Section 5 of the GNU AGPL version 3.
-//
-// Pursuant to Section 7(b) of the License you must retain the original Product logo when
-// distributing the program. Pursuant to Section 7(e) we decline to grant you any rights under
-// trademark law for use of our trademarks.
-//
-// All the Product's GUI elements, including illustrations and icon sets, as well as technical writing
-// content are licensed under the terms of the Creative Commons Attribution-ShareAlike 4.0
-// International. See the License terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
-
-using ASC.EventBus.Exceptions;
-
-namespace ASC.Data.Backup.IntegrationEvents.EventHandling;
-
-[Scope]
-public class BackupRequestIntegrationEventHandler : IIntegrationEventHandler<BackupRequestIntegrationEvent>
-{
-    private readonly BackupAjaxHandler _backupAjaxHandler;
-    private readonly ILog _logger;
-    private readonly TenantManager _tenantManager;
-    private readonly SecurityContext _securityContext;
-    private readonly AuthManager _authManager;
-    private readonly BackupWorker _backupWorker;
-
-    public BackupRequestIntegrationEventHandler(
-        BackupAjaxHandler backupAjaxHandler,
-        IOptionsMonitor<ILog> logger,
-        TenantManager tenantManager,
-        SecurityContext securityContext,
-        AuthManager authManager,
-        BackupWorker backupWorker)
-    {
-        _tenantManager = tenantManager;
-        _authManager = authManager;
-        _securityContext = securityContext;
-        _backupAjaxHandler = backupAjaxHandler;
-        _logger = logger.CurrentValue;
-        _backupWorker = backupWorker;
-    }
-
-    public async Task Handle(BackupRequestIntegrationEvent @event)
-    {
-        _logger.InfoFormat("----- Handling integration event: {IntegrationEventId} at {AppName} - ({@IntegrationEvent})", @event.Id, Program.AppName, @event);
-
-        if (!@event.Redelivered)
-        {
-            if (_backupWorker.IsInstanceTooBusy())
-            {
-                throw new IntegrationEventRejectExeption(@event.Id);
-            }
-        }
-
-        _tenantManager.SetCurrentTenant(@event.TenantId);
-        _securityContext.AuthenticateMeWithoutCookie(_authManager.GetAccountByID(@event.TenantId, @event.CreateBy));
-
-        if (@event.IsScheduled)
-        {
-            _backupWorker.StartScheduledBackup(new EF.Model.BackupSchedule
-            {
-                BackupMail = @event.BackupMail,
-                BackupsStored = @event.BackupsStored,
-                StorageBasePath = @event.StorageBasePath,
-                StorageParams = JsonConvert.SerializeObject(@event.StorageParams),
-                StorageType = @event.StorageType,
-                TenantId = @event.TenantId
-            });
-        }
-        else
-        {
-            _backupAjaxHandler.StartBackup(@event.StorageType, @event.StorageParams, @event.BackupMail);
-        }
-
-        await Task.CompletedTask;
-    }
-}
